Improve error checks for set operations
Before this commit, ErrorIfCannotPushdownSubquery() was failing to check
the subqueries within the subquerues that reside in the set operations.
Similarly, Citus was only considering the top level children of the set
operations and not recursing into the subqueries within those children.
This commit adds the logic to recursively traverse into the union
queries.